requirements.txt
   streamlit
   PyMuPDF
   requests
   chromadb
   sentence-transformers
   pyyaml